Fixed powl to pow in UUID random distribution
[C] [W]

+++ Peach/src/Peach/System/UUID.cpp
@@ -7,7 +7,7 @@ namespace Peach
 {
 	static std::random_device s_RandomDevice;
 	static std::default_random_engine s_Engine(s_RandomDevice());
-	static std::uniform_int_distribution<size_t> s_UniformDistribution(powl(10, 19));
+	static std::uniform_int_distribution<size_t> s_UniformDistribution(pow(10, 19));
 
 	UUID::UUID()
 		: m_UUID(s_UniformDistribution(s_Engine))
